# contracts/openapi/openapi.v1.yaml
openapi: 3.0.3
info:
  title: Lpsaring API Contract
  version: 1.0.0
  description: |
    Source of truth kontrak API domain prioritas (Auth, Profile, Devices, Transactions).
    Ruang lingkup ini sengaja fokus untuk mengurangi drift backend-frontend pada flow kritikal.
    Revisi internal 2026-02-26: refactor wiring transaksi/payment tanpa perubahan signature endpoint.
servers:
  - url: /api
    description: Relative API base path
...
paths:
...
  /users/check-or-register:
    post:
      tags: [Profile]
      summary: Cek status registrasi nomor telepon (publik, tanpa autentikasi)
      description: |
        Dipakai halaman pembelian (`pages/beli/index.vue`) untuk memverifikasi apakah
        nomor telepon sudah terdaftar sebelum melanjutkan flow OTP / pendaftaran.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneCheckRequest'
      responses:
        '200':
          description: Hasil pengecekan nomor telepon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneCheckResponse'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '422':
          $ref: '#/components/responses/ErrorValidation'
        '500':
          $ref: '#/components/responses/ErrorInternal'
  /users/database-update-submissions:
    post:
      tags: [Profile]
      summary: Submit pemutakhiran data pengguna (publik, tanpa autentikasi)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicDatabaseUpdateSubmissionRequest'
      responses:
        '201':
          description: Submission berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '422':
          $ref: '#/components/responses/ErrorValidation'
  /users/database-update-submissions/status:
    get:
      tags: [Profile]
      summary: Cek status pemutakhiran data berdasarkan nomor HP (publik, tanpa autentikasi)
      parameters:
        - in: query
          name: phone
          required: true
          schema:
            type: string
          description: Nomor HP yang digunakan saat import (format bebas, dinormalisasi server)
      responses:
        '200':
          description: Status pemutakhiran data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUpdateSubmissionStatusResponse'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  responses:
    ErrorBadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ErrorForbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ErrorValidation:
      description: Validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  details:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidationErrorDetail'
    ErrorInternal:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code
          example: AUTH_UNAUTHORIZED
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDetail'
        request_id:
          type: string
    ValidationErrorDetail:
      type: object
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
...
    PhoneCheckRequest:
      type: object
      required: [phone_number]
      properties:
        phone_number:
          type: string
          description: Nomor HP (format bebas; dinormalisasi server ke E.164)
        full_name:
          type: string
          nullable: true
          description: Opsional; bila user sudah ada tapi full_name kosong, akan diisi
    PhoneCheckResponse:
      type: object
      required: [user_exists]
      properties:
        user_exists:
          type: boolean
        user_id:
          type: string
          format: uuid
          nullable: true
        message:
          type: string
          nullable: true
    PublicDatabaseUpdateSubmissionRequest:
      type: object
      required: [full_name, role, phone_number]
      properties:
        full_name:
          type: string
        role:
          type: string
          enum: [USER, KOMANDAN, TAMPING]
        blok:
          type: string
          nullable: true
        kamar:
          type: string
          nullable: true
        tamping_type:
          type: string
          nullable: true
        phone_number:
          type: string
    PublicUpdateSubmissionStatusResponse:
      type: object
      required: [success, status]
      properties:
        success:
          type: boolean
        status:
          type: string
          enum: [none, reviewing, approved]
...